Add files via upload
This Python project is a To-Do List Application that allows users to manage tasks interactively through a command-line interface. Users can:

View all tasks: See a list of tasks with their current status.
Add a new task: Input and add new tasks to the list.
Mark tasks as done: Update the status of tasks to "Done."
Update a task: Edit the description of existing tasks.
Remove a task: Delete tasks from the list.
The program repeatedly displays a menu and performs actions based on user input until the user decides to exit.






This Python project is a To-Do List Application that allows users to manage tasks interactively through a command-line interface. Users can:

View all tasks: See a list of tasks with their current status.
Add a new task: Input and add new tasks to the list.
Mark tasks as done: Update the status of tasks to "Done."
Update a task: Edit the description of existing tasks.
Remove a task: Delete tasks from the list.
The program repeatedly displays a menu and performs actions based on user input until the user decides to exit.